requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.5b1
click==8.0.0
colorama==0.4.4
Django==3.2
django-appconf==1.0.4
django-debug-toolbar==3.2.1
django-imagekit==4.0.2
flake8==3.9.2
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
pilkit==2.0
Pillow==8.2.0
pycodestyle==2.7.0
pyflakes==2.3.1
python-decouple==3.4
pytz==2021.1
regex==2021.4.4
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
